requirements.txt
psycopg2-binary==2.9.11
requests==2.32.5
autogen-agentchat @ git+https://github.com/patrickstolc/autogen.git@0.2#egg=autogen-agentchat
autogen==0.3.1
mistralai==1.2.3
ollama==0.3.3
fix-busted-json==0.0.18